services/events_etl/src/requirements.txt
backoff==2.2.1
clickhouse-driver==0.2.9
kafka-python==2.2.10
pydantic-settings==2.9.1
python-dateutil==2.9.0.post0